requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
pydantic==2.8.2
sqlalchemy==2.0.32
mysql-connector-python==9.0.0
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
email-validator==2.2.0
httptools==0.6.1
uvloop==0.20.0
